[Tool] separat functions from makeTopiQL.js
- create tools folder to hold functions
- migrate functions from makeTopiQL to graphqlSchemaTool
- add / edit / omit comments
